refactor: dashboard layout and components for clarity
