+++ apps/web/src/app/[locale]/(dashboard)/(home)/page.tsx
