+++ apps/web/src/app/[locale]/(dashboard)/directories/new/new-directory-client.tsx
@@ -24,30 +24,11 @@ export default function NewDirectoryClient({
 
     if (creationMode === null) {
         return (
-            <div className="max-w-5xl mx-auto px-4 py-8">
+            <div className="max-w-5xl mx-auto px-4 sm:px-6 lg:px-8 py-6 lg:py-8">
                 {/* GitHub Connection Alert */}
                 <GitHubConnectionAlert githubConnected={githubConnected} />
 
                 <div className="mb-8">
-                    <button
-                        onClick={() => router.back()}
-                        className="flex items-center gap-2 text-text-secondary dark:text-text-secondary-dark hover:text-text dark:hover:text-text-dark transition-colors mb-4"
-                    >
-                        <svg
-                            className="w-5 h-5"
-                            fill="none"
-                            stroke="currentColor"
-                            viewBox="0 0 24 24"
-                        >
-                            <path
-                                strokeLinecap="round"
-                                strokeLinejoin="round"
-                                strokeWidth={2}
-                                d="M10 19l-7-7m0 0l7-7m-7 7h18"
-                            />
-                        </svg>
-                        Back
-                    </button>
                     <h1 className="text-3xl font-bold text-text dark:text-text-dark">
                         Create New Directory
                     </h1>
@@ -177,7 +158,7 @@ export default function NewDirectoryClient({
     }
 
     return (
-        <div className="flex gap-6 max-w-7xl mx-auto px-4 py-8">
+        <div className="flex gap-6 max-w-7xl mx-auto px-4 sm:px-6 lg:px-8 py-6 lg:py-8">
             {/* Main Content */}
             <div className="flex-1">
                 <div className="mb-8">
