+++ apps/web/src/app/[locale]/(dashboard)/layout.tsx
@@ -1,3 +1,14 @@
-export default function DashboardLayout({ children }: { children: React.ReactNode }) {
-    return <>{children}</>;
+import { getAuthUser } from '@/lib/auth';
+import { DashboardLayoutClient } from './layout-client';
+
+export default async function DashboardLayout({ children }: { children: React.ReactNode }) {
+    const user = await getAuthUser();
+
+    if (!user) {
+        // This shouldn't happen as middleware should redirect
+        // but keeping as a safety check
+        return null;
+    }
+
+    return <DashboardLayoutClient user={user}>{children}</DashboardLayoutClient>;
 }
